fix size_t

+++ src/Interface/LoggerInterface.h
@@ -37,7 +37,7 @@ namespace Mengine
 		virtual bool validMessage( EMessageLevel _level, uint32_t _flag ) const = 0;
 
 	public:
-		virtual void log( EMessageLevel _level, uint32_t _flag, const Char * _data, size_t _size ) = 0;
+		virtual void log( EMessageLevel _level, uint32_t _flag, const Char * _data, uint32_t _size ) = 0;
 		virtual void flush() = 0;
 	};
     //////////////////////////////////////////////////////////////////////////
